v1 of Advanced R
Only comitting selected chapters. Using this
mainly as a good foundational review of R; but not going deep into things like function factories, OOP, or meta-programming.

This has been excellent, however, for getting to grips with the underlying structures and ubiquitous S3 objects in R, and should make it easier to predict behavior and investigate documentation.